Move refreshAllDatabases back to Explorer
